backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
mysql-connector-python==8.3.0
pymysql==1.1.0
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
yfinance==0.2.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9